+++ planner_cspace/src/rotation_cache.cpp
@@ -33,7 +33,6 @@
 #include <vector>
 
 #include <planner_cspace/cyclic_vec.h>
-
 #include <planner_cspace/planner_3d/rotation_cache.h>
 
 void RotationCache::Page::reset(const CyclicVecInt<3, 2>& size)
@@ -81,7 +80,7 @@ void RotationCache::reset(
           const float cos_v = std::cos(v[2]);
           const float r1 = v[1] + v[0] * cos_v / sin_v;
           const float r2 = std::copysign(
-              sqrtf(powf(v[0], 2.0) + powf(v[0] * cos_v / sin_v, 2.0)),
+              std::sqrt(std::pow(v[0], 2) + std::pow(v[0] * cos_v / sin_v, 2)),
               v[0] * sin_v);
           r.radiuses(d) = std::pair<float, float>(r1, r2);
         }
