+++ trajectory_tracker/src/trajectory_recorder.cpp
@@ -35,15 +35,15 @@
    This software was implemented to accomplish the above research.
  */
 
+#include <cmath>
+#include <string>
+
 #include <ros/ros.h>
 
 #include <geometry_msgs/Twist.h>
 #include <nav_msgs/Path.h>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 #include <tf2_ros/transform_listener.h>
-
-#include <math.h>
-#include <string>
 
 #include <neonavigation_common/compatibility.h>
 
@@ -94,7 +94,7 @@ RecorderNode::~RecorderNode()
 
 float dist2d(geometry_msgs::Point& a, geometry_msgs::Point& b)
 {
-  return sqrtf(powf(a.x - b.x, 2) + powf(a.y - b.y, 2));
+  return std::sqrt(std::pow(a.x - b.x, 2) + std::pow(a.y - b.y, 2));
 }
 
 void RecorderNode::spin()
